requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.19.0
python-decouple==3.6
simplejson==3.17.6
tomli==2.0.1
typing_extensions==4.3.0
yarl==1.8.1
zipp==3.8.1
